Write code to view other users' root folders
Clicking username in sidebar shows contents of their root folder
