+++ assets/scripts/templates/user.handlebars
@@ -1,5 +1,5 @@
 {{#each users}}
 
-  <li><a href="#" data-path=",{{_id}}">{{email}}</a></li>
+  <li><a class="username" href="#" data-path=",{{_id}}">{{email}}</a></li>
 
 {{/each}}
