+++ assets/scripts/ui.js
@@ -20,36 +20,10 @@ const createSuccess = () => {
   console.log('folder created');
 };
 
-const displayUserFolder = function(data){
-  console.log(data);
-  let userFolderTemplate = require('./templates/current-user-folder.handlebars');
-    $('#main-content').html(userFolderTemplate({
-      folders: data.folders
-    }));
-};
-
-const displayUserFile = function(data){
-  console.log(data);
-  let userFileTemplate = require('./templates/current-user-file.handlebars');
-    $('#main-content').append(userFileTemplate({
-      files: data.files
-    }));
-};
-
-const displayUsers = function(data){
-  let userTemplate = require('./templates/user.handlebars');
-  $('.sidebar-nav').html(userTemplate({
-      users: data.users,
-    }));
-};
-
 module.exports = {
   failure,
   success,
   signOutSuccess,
   app,
   createSuccess,
-  displayUsers,
-  displayUserFolder,
-  displayUserFile,
 };
